requirements.txt
aiofiles==0.6.0
APScheduler==3.8.1
asgiref==3.4.1
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.4
dnspython==2.0.0
ecdsa==0.14.1
email-validator==1.1.3
fake-useragent==0.1.11
fastapi==0.68.1
h11==0.11.0
idna==2.10
imbox==0.9.8
importlib-metadata==3.3.0
iniconfig==1.1.1
lxml==4.6.5
MarkupSafe==1.1.1
multidict==5.2.0
outcome==1.1.0
packaging==20.8
passlib==1.7.4
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pydub==0.25.1
pyOpenSSL==21.0.0
pyparsing==3.0.6
pytest==6.2.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
qrcode==7.3
ratelimiter==1.2.0.post0
requests==2.25.1
rsa==4.6
selenium==4.1.0
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
SQLAlchemy==1.3.22
starlette==0.14.2
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.7.4.3
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
uvicorn==0.15.0
wsproto==1.0.0
yarl==1.7.2
zipp==3.4.0
python-dateutil==2.8.2
ltp==4.1.4
cryptography=1.3.5